backend/requirements.txt
Django==4.1.3
django-colorfield==0.7.2
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.14.0
djoser==2.1.0
drf-yasg==1.21.4
flake8==6.0.0
gunicorn==20.1.0
Pillow==9.3.0
psycopg2-binary==2.9.5
python-dotenv==0.21.0